+++ CHANGELOG.md
@@ -3,6 +3,21 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.0.0-alpha.41](https://github.com/d-zero-dev/linters/compare/v5.0.0-alpha.40...v5.0.0-alpha.41) (2024-07-31)
+
+### Bug Fixes
+
+- **deps:** update dependency cz-customizable to v7.2.1 ([1e91c96](https://github.com/d-zero-dev/linters/commit/1e91c968dd919aeaa587c89d4f25552de5fb07e6))
+- **deps:** update dependency eslint-plugin-jsdoc to v48.10.1 ([830678a](https://github.com/d-zero-dev/linters/commit/830678af68d2ca2a64f66714d822e54d3fc69b7b))
+- **deps:** update dependency eslint-plugin-jsdoc to v48.9.2 ([56d3efb](https://github.com/d-zero-dev/linters/commit/56d3efb291ac790c7e34f8534c7e8ae5fc789ac3))
+- **deps:** update dependency stylelint to v16.8.0 ([8827315](https://github.com/d-zero-dev/linters/commit/8827315eded99568de6f6a1ad7c90ab85918c5a7))
+- **deps:** update typescript-eslint monorepo to v7.18.0 ([39a509c](https://github.com/d-zero-dev/linters/commit/39a509c5ccad510d8cd8282bc9b82e79da2e1bee))
+
+### Features
+
+- **lint-staged:** export type declaration files ([349a811](https://github.com/d-zero-dev/linters/commit/349a811ee74f143f669b6e95499463b5a0a5e8c8))
+- **lint-staged:** 特定のファイルに対しての特定のリンターの実行を除外するオプションを追加 ([fea733b](https://github.com/d-zero-dev/linters/commit/fea733b452740c6fa68f3b5ad73b9ef0cf1aa6ec))
+
 # [5.0.0-alpha.40](https://github.com/d-zero-dev/linters/compare/v5.0.0-alpha.39...v5.0.0-alpha.40) (2024-07-17)
 
 ### Bug Fixes